refactor: change distance unit type from Int to Double in CityStatus and LocationStatus

+++ app/src/main/java/il/soulSalttrader/shabbattimes/content/city/CityStatus.kt
@@ -3,5 +3,5 @@ package il.soulSalttrader.shabbattimes.content.city
 sealed interface CityStatus {
     data object Unknown : CityStatus
     data object Current : CityStatus
-    data class Nearby(val km: Int) : CityStatus
+    data class Nearby(val km: Double) : CityStatus
 }
