+++ app/src/main/java/il/soulSalttrader/shabbattimes/location/LocationStatus.kt
@@ -3,5 +3,5 @@ package il.soulSalttrader.shabbattimes.location
 sealed interface LocationStatus {
     data object Unknown : LocationStatus
     data object Current : LocationStatus
-    data class Nearby(val km: Int) : LocationStatus
+    data class Nearby(val km: Double) : LocationStatus
 }
